Fixed the JPA errors, tested services and added Rest-layer with DTO's.

+++ dao/src/main/java/nl/luukhermans/dao/UserDao.java
@@ -10,7 +10,7 @@ public interface UserDao {
 
     Collection<User> getAllUsers();
 
-    User findByID(int ID);
+    User findByID(Long ID);
 
     User findByUsername(String username);
